feature/MTCE-254-Monorepo-v6-refactor-tests-structure
https://eonx.atlassian.net/browse/MTCE-254
