+++ composer.json
@@ -176,20 +176,19 @@
     "autoload-dev": {
         "psr-4": {
             "EonX\\EasyActivity\\Tests\\": "packages/EasyActivity/tests",
+            "EonX\\EasyActivity\\Tests\\Application\\": "packages/EasyActivity/tests/Application/src",
             "EonX\\EasyActivity\\Tests\\Fixture\\App\\": "packages/EasyActivity/tests/Fixture/app/src",
-            "EonX\\EasyActivity\\Tests\\Unit\\": "packages/EasyActivity/tests/Unit/src",
             "EonX\\EasyApiPlatform\\Tests\\": "packages/EasyApiPlatform/tests",
             "EonX\\EasyApiPlatform\\Tests\\Application\\": "packages/EasyApiPlatform/tests/Application/src",
             "EonX\\EasyApiPlatform\\Tests\\Fixture\\App\\": "packages/EasyApiPlatform/tests/Fixture/app/src",
-            "EonX\\EasyApiPlatform\\Tests\\Unit\\": "packages/EasyApiPlatform/tests/Unit/src",
             "EonX\\EasyApiToken\\Tests\\": "packages/EasyApiToken/tests",
             "EonX\\EasyApiToken\\Tests\\Unit\\": "packages/EasyApiToken/tests/Unit/src",
             "EonX\\EasyApiToken\\Tests\\Unit\\Bundle\\": "packages/EasyApiToken/tests/Unit/bundle",
             "EonX\\EasyApiToken\\Tests\\Unit\\Laravel\\": "packages/EasyApiToken/tests/Unit/laravel",
             "EonX\\EasyAsync\\Tests\\": "packages/EasyAsync/tests",
+            "EonX\\EasyAsync\\Tests\\Application\\": "packages/EasyAsync/tests/Application/src",
+            "EonX\\EasyAsync\\Tests\\Application\\Bundle\\": "packages/EasyAsync/tests/Application/bundle",
             "EonX\\EasyAsync\\Tests\\Fixture\\App\\": "packages/EasyAsync/tests/Fixture/app/src",
-            "EonX\\EasyAsync\\Tests\\Unit\\": "packages/EasyAsync/tests/Unit/src",
-            "EonX\\EasyAsync\\Tests\\Unit\\Bundle\\": "packages/EasyAsync/tests/Unit/bundle",
             "EonX\\EasyBankFiles\\Tests\\": "packages/EasyBankFiles/tests",
             "EonX\\EasyBankFiles\\Tests\\Unit\\": "packages/EasyBankFiles/tests/Unit/src",
             "EonX\\EasyBatch\\Tests\\": "packages/EasyBatch/tests",
