+++ packages/EasyActivity/tests/Application/AbstractApplicationTestCase.php
@@ -1,17 +1,20 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Tests\Unit;
+namespace EonX\EasyActivity\Tests\Application;
 
-use EonX\EasyApiPlatform\Tests\Fixture\App\Kernel\ApplicationKernel;
+use Doctrine\ORM\Tools\SchemaTool;
+use EonX\EasyActivity\Tests\Fixture\App\Kernel\ApplicationKernel;
 use EonX\EasyTest\Common\Trait\ContainerServiceTrait;
+use EonX\EasyTest\Common\Trait\DatabaseEntityTrait;
 use EonX\EasyTest\Common\Trait\PrivatePropertyAccessTrait;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Filesystem\Filesystem;
 
-abstract class AbstractUnitTestCase extends KernelTestCase
+abstract class AbstractApplicationTestCase extends KernelTestCase
 {
     use ContainerServiceTrait;
+    use DatabaseEntityTrait;
     use PrivatePropertyAccessTrait;
 
     public static function tearDownAfterClass(): void
@@ -29,5 +32,15 @@ abstract class AbstractUnitTestCase extends KernelTestCase
     protected static function getKernelClass(): string
     {
         return ApplicationKernel::class;
+    }
+
+    protected function initDatabase(): void
+    {
+        $entityManager = self::getEntityManager();
+        $metaData = $entityManager->getMetadataFactory()
+            ->getAllMetadata();
+        $schemaTool = new SchemaTool($entityManager);
+        $schemaTool->dropSchema($metaData);
+        $schemaTool->updateSchema($metaData);
     }
 }
