+++ packages/EasyActivity/tests/Fixture/app/config/packages/default_subject_config/easy_activity.php
@@ -7,10 +7,10 @@ use EonX\EasyActivity\Tests\Fixture\App\Entity\Article;
 use Symfony\Config\EasyActivityConfig;
 
 /**
- * @see \EonX\EasyActivity\Tests\Unit\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceeds
- * @see \EonX\EasyActivity\Tests\Unit\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceedsWithCollections
- * @see \EonX\EasyActivity\Tests\Unit\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceedsWithRelatedObjects
- * @see \EonX\EasyActivity\Tests\Unit\EasyDoctrine\Subscriber\EasyDoctrineEntityEventsSubscriberTest::testLoggerSucceedsForDeletedSubjects
+ * @see \Application\src\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceeds
+ * @see \Application\src\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceedsWithCollections
+ * @see \Application\src\Common\Factory\ActivityLogEntryFactoryTest::testCreateSucceedsWithRelatedObjects
+ * @see \Application\src\EasyDoctrine\Subscriber\EasyDoctrineEntityEventsSubscriberTest::testLoggerSucceedsForDeletedSubjects
  */
 return static function (EasyActivityConfig $easyActivityConfig): void {
     $easyActivityConfig->subjects(Article::class);
