+++ packages/EasyActivity/tests/Fixture/app/config/packages/only_disallowed_properties/easy_activity.php
@@ -7,7 +7,7 @@ use EonX\EasyActivity\Tests\Fixture\App\Entity\Article;
 use Symfony\Config\EasyActivityConfig;
 
 /**
- * @see \EonX\EasyActivity\Tests\Unit\EasyDoctrine\Subscriber\EasyDoctrineEntityEventsSubscriberTest::provideProperties
+ * @see \Application\src\EasyDoctrine\Subscriber\EasyDoctrineEntityEventsSubscriberTest::provideProperties
  */
 return static function (EasyActivityConfig $easyActivityConfig): void {
     $easyActivityConfig->subjects(Article::class)
