+++ packages/EasyAsync/composer.json
@@ -35,9 +35,9 @@
     "autoload-dev": {
         "psr-4": {
             "EonX\\EasyAsync\\Tests\\": "tests",
-            "EonX\\EasyAsync\\Tests\\Fixture\\App\\": "tests/Fixture/app/src",
-            "EonX\\EasyAsync\\Tests\\Unit\\": "tests/Unit/src",
-            "EonX\\EasyAsync\\Tests\\Unit\\Bundle\\": "tests/Unit/bundle"
+            "EonX\\EasyAsync\\Tests\\Application\\": "tests/Application/src",
+            "EonX\\EasyAsync\\Tests\\Application\\Bundle\\": "tests/Application/bundle",
+            "EonX\\EasyAsync\\Tests\\Fixture\\App\\": "tests/Fixture/app/src"
         }
     },
     "suggest": {
