+++ packages/EasyAsync/tests/Fixture/app/config/packages/messages_and_time_limits/easy_async.php
@@ -6,7 +6,7 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 use Symfony\Config\EasyAsyncConfig;
 
 /**
- * @see \EonX\EasyAsync\Tests\Unit\Bundle\EasyAsyncBundleTest::testMessengerConfigWithMessagesAndTimeLimits
+ * @see \EonX\EasyAsync\Tests\Application\Bundle\EasyAsyncBundleTest::testMessengerConfigWithMessagesAndTimeLimits
  */
 return static function (EasyAsyncConfig $easyAsyncConfig): void {
     $messengerWorkerConfig = $easyAsyncConfig->messengerWorker();
