+++ packages/EasyAsync/tests/Fixture/app/config/packages/messages_limit/easy_async.php
@@ -6,7 +6,7 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 use Symfony\Config\EasyAsyncConfig;
 
 /**
- * @see \EonX\EasyAsync\Tests\Unit\Bundle\EasyAsyncBundleTest::testMessengerConfigWithMessagesLimit
+ * @see \EonX\EasyAsync\Tests\Application\Bundle\EasyAsyncBundleTest::testMessengerConfigWithMessagesLimit
  */
 return static function (EasyAsyncConfig $easyAsyncConfig): void {
     $messengerWorkerConfig = $easyAsyncConfig->messengerWorker();
